+++ lib/powersync.dart
@@ -1,12 +1,12 @@
 /// PowerSync Dart SDK.
 ///
 /// Use [PowerSyncDatabase] to open a database.
-library powersync;
+library;
 
 export 'src/powersync_database.dart';
 export 'src/schema.dart';
-export 'src/sqlite_connection.dart';
 export 'src/connector.dart';
 export 'src/crud.dart';
 export 'src/streaming_sync.dart' show SyncStatus;
 export 'src/uuid.dart';
+export 'src/open_factory.dart';