+++ lib/src/database_utils.dart
@@ -1,43 +1,6 @@
 import 'dart:async';
 
-import 'package:powersync/powersync.dart';
-import 'package:sqlite3/sqlite3.dart' as sqlite;
-
-Future<T> internalReadTransaction<T>(SqliteReadContext ctx,
-    Future<T> Function(SqliteReadContext tx) callback) async {
-  try {
-    await ctx.getAll('BEGIN');
-    final result = await callback(ctx);
-    await ctx.getAll('END TRANSACTION');
-    return result;
-  } catch (e) {
-    try {
-      await ctx.getAll('ROLLBACK');
-    } catch (e) {
-      // In rare cases, a ROLLBACK may fail.
-      // Safe to ignore.
-    }
-    rethrow;
-  }
-}
-
-Future<T> internalWriteTransaction<T>(SqliteWriteContext ctx,
-    Future<T> Function(SqliteWriteContext tx) callback) async {
-  try {
-    await ctx.execute('BEGIN IMMEDIATE');
-    final result = await callback(ctx);
-    await ctx.execute('COMMIT');
-    return result;
-  } catch (e) {
-    try {
-      await ctx.execute('ROLLBACK');
-    } catch (e) {
-      // In rare cases, a ROLLBACK may fail.
-      // Safe to ignore.
-    }
-    rethrow;
-  }
-}
+import 'package:sqlite_async/sqlite3.dart' as sqlite;
 
 Future<T> asyncDirectTransaction<T>(sqlite.Database db,
     FutureOr<T> Function(sqlite.Database db) callback) async {
@@ -70,19 +33,4 @@ Future<T> asyncDirectTransaction<T>(sqlite.Database db,
     }
   }
   throw AssertionError('Should not reach this');
-}
-
-/// Given a SELECT query, return the tables that the query depends on.
-Future<Set<String>> getSourceTables(SqliteReadContext ctx, String sql) async {
-  final rows = await ctx.getAll('EXPLAIN QUERY PLAN $sql');
-  Set<String> tables = {};
-  final re = RegExp(r'^(SCAN|SEARCH)( TABLE)? (.+?)( USING .+)?$');
-  for (var row in rows) {
-    final detail = row['detail'];
-    final match = re.firstMatch(detail);
-    if (match != null) {
-      tables.add(match.group(3)!);
-    }
-  }
-  return tables;
 }