+++ test/crud_test.dart
@@ -1,5 +1,5 @@
 import 'package:powersync/powersync.dart';
-import 'package:sqlite3/sqlite3.dart' as sqlite;
+import 'package:sqlite_async/sqlite3.dart' as sqlite;
 import 'package:test/test.dart';
 
 import 'util.dart';